make annotations more clear in args with list of values
